requirement.txt
GitPython==3.1.31
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
joblib==1.2.0
numpy==1.21.6
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sacremoses==0.0.53
scipy==1.7.3
setuptools==67.7.2
six==1.16.0
smmap==5.0.0
tokenizers==0.13.3
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.65.0
transformers==4.16.2
tree-sitter==0.20.1
typing_extensions==4.5.0
urllib3==2.0.2
wheel==0.40.0
zipp==3.15.0
